Add tooltips for keyboard navigation
