+++ package.json
@@ -78,6 +78,7 @@
         "react-modal": "^3.10.1",
         "react-router": "^5.0.1",
         "react-router-dom": "^5.0.1",
+        "react-tooltip": "^3.10.0",
         "request": "^2.88.0",
         "reset-css": "^4.0.1",
         "snyk": "^1.221.1",
