--- src/components/widgets/Navigasjonsknapper.css
@@ -1,3 +1,0 @@
-.Navigasjonsknapper .knapp:not(:last-child) {
-    margin-right: 0.5rem;
-}
