+++ src/components/widgets/Navigasjonsknapper/index.js
@@ -1,0 +1,3 @@
+import Navigasjonsknapper from './Navigasjonsknapper';
+
+export default Navigasjonsknapper;
